move fns in ampdoc.js to services.js and delete
