+++ ads/google/a4a/test/test-traffic-experiments.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {ampdocServiceFor} from '../../../../src/ampdoc';
+import {ampdocServiceFor} from '../../../../src/services';
 import {installDocService} from '../../../../src/service/ampdoc-impl';
 import {
   addExperimentIdToElement,
