+++ extensions/amp-dynamic-css-classes/0.1/test/test-runtime-classes.js
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-import {ampdocServiceFor} from '../../../../src/ampdoc';
+import {ampdocServiceFor, viewerForDoc} from '../../../../src/services';
 import {installDocService} from '../../../../src/service/ampdoc-impl';
 import {installPlatformService} from '../../../../src/service/platform-impl';
 import {installViewerServiceForDoc} from '../../../../src/service/viewer-impl';
-import {viewerForDoc} from '../../../../src/services';
 import {vsyncForTesting} from '../../../../src/service/vsync-impl';
 import {installDynamicClassesForTesting} from '../amp-dynamic-css-classes';
