+++ extensions/amp-fx-parallax/0.1/amp-fx-parallax.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {ampdocServiceFor} from '../../../src/ampdoc';
+import {ampdocServiceFor} from '../../../src/services';
 import {installParallaxForDoc} from '../../../src/service/parallax-impl';
 import {onDocumentReady} from '../../../src/document-ready';
