+++ src/service/vsync-impl.js
@@ -15,14 +15,12 @@
  */
 
 import {Pass} from '../pass';
-import {ampdocServiceFor} from '../ampdoc';
+import {ampdocServiceFor, viewerForDoc, viewerPromiseForDoc} from '../services';
 import {cancellation} from '../error';
 import {dev, rethrowAsync} from '../log';
 import {documentStateFor} from './document-state';
-
 import {registerServiceBuilder, getService} from '../service';
 import {installTimerService} from './timer-impl';
-import {viewerForDoc, viewerPromiseForDoc} from '../services';
 import {JankMeter} from './jank-meter';
 
 /** @const {time} */
