+++ testing/describes.js
@@ -90,7 +90,11 @@ import {
 } from './fake-dom';
 import {installFriendlyIframeEmbed} from '../src/friendly-iframe-embed';
 import {doNotLoadExternalResourcesInTest} from './iframe';
-import {ampdocServiceFor} from '../src/ampdoc';
+import {
+  ampdocServiceFor,
+  extensionsFor,
+  resourcesForDoc,
+} from '../src/services';
 import {
   adopt,
   adoptShadowMode,
@@ -108,7 +112,6 @@ import {
   installExtensionsService,
   registerExtension,
 } from '../src/service/extensions-impl';
-import {extensionsFor, resourcesForDoc} from '../src/services';
 import {resetScheduledElementForTesting} from '../src/custom-element';
 import {setStyles} from '../src/style';
 import * as sinon from 'sinon';
