+++ testing/iframe.js
@@ -18,7 +18,7 @@ import {AmpEvents} from '../src/amp-events';
 import {BindEvents} from '../extensions/amp-bind/0.1/bind-events';
 import {FakeLocation} from './fake-dom';
 import {FormEvents} from '../extensions/amp-form/0.1/form-events';
-import {ampdocServiceFor} from '../src/ampdoc';
+import {ampdocServiceFor, resourcesForDoc} from '../src/services';
 import {cssText} from '../build/css';
 import {deserializeMessage, isAmpMessage} from '../src/3p-frame-messaging';
 import {parseIfNeeded} from '../src/iframe-helper';
@@ -32,7 +32,6 @@ import installCustomElements from
 import {installDocService} from '../src/service/ampdoc-impl';
 import {installExtensionsService} from '../src/service/extensions-impl';
 import {installStyles} from '../src/style-installer';
-import {resourcesForDoc} from '../src/services';
 
 let iframeCount = 0;
